Move plot drawing to a separate file
